Refactor so that close button inserts latex
editor_plugin.js now queries the popup for latex, rather than the popup
pushing the latex back to the main window. This change was necessitated
by the need to make use of the onClose handler (so that we insert the
latex even when the user hits the X button), which is only available
from the main window.