Reopen the dialog when user clicks a mathquill img
Added a local variable editing that acts as a lock so that the
onNodeChange doesn't infinite loop (it gets called again after we try
to open the window, for some godforsaken reason only the creators of
tinyMCE understand).

Added a class "rendered-latex" to the image so that we can find it
again in onNodeChange.